Don't use container directory
