--- zipkin-lens/src/containers/GlobalSearch/GlobalSearchContainer.js
@@ -1,36 +1,0 @@
-/*
- * Copyright 2015-2019 The OpenZipkin Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-import { connect } from 'react-redux';
-
-import GlobalSearch from '../../components/GlobalSearch/GlobalSearch';
-import { fetchTraces } from '../../actions/traces-action';
-import { fetchServices } from '../../actions/services-action';
-
-const mapStateToProps = state => ({
-  conditions: state.globalSearch.conditions,
-  lookbackCondition: state.globalSearch.lookbackCondition,
-  limitCondition: state.globalSearch.limitCondition,
-});
-
-const mapDispatchToProps = dispatch => ({
-  fetchTraces: params => dispatch(fetchTraces(params)),
-  fetchServices: () => dispatch(fetchServices()),
-});
-
-const GlobalSearchContainer = connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(GlobalSearch);
-
-export default GlobalSearchContainer;
